Move FilteredTrackDataset into its own module
