+++ OTAnalytics/plugin_datastore/python_track_store.py
@@ -4,6 +4,7 @@ from datetime import datetime
 from math import ceil
 from typing import Callable, Iterable, Optional, Sequence
 
+from domain.track_dataset.filtered_track_dataset import FilterByClassTrackDataset
 from more_itertools import batched
 
 from OTAnalytics.application.logger import logger
@@ -33,7 +34,6 @@ from OTAnalytics.domain.track_dataset.track_dataset import (
     START_X,
     START_Y,
     TRACK_GEOMETRY_FACTORY,
-    FilterByClassTrackDataset,
     IntersectionPoint,
     TrackDataset,
     TrackDoesNotExistError,
