+++ tests/OTAnalytics/domain/track_dataset/test_filtered_track_dataset.py
@@ -3,13 +3,11 @@ from unittest.mock import Mock
 
 import pytest
 from _pytest.fixtures import FixtureRequest
+from domain.track_dataset.filtered_track_dataset import FilterByClassTrackDataset
 
 from OTAnalytics.domain.otc_classes import OtcClasses
 from OTAnalytics.domain.track import Track, TrackId
-from OTAnalytics.domain.track_dataset.track_dataset import (
-    FilterByClassTrackDataset,
-    TrackDoesNotExistError,
-)
+from OTAnalytics.domain.track_dataset.track_dataset import TrackDoesNotExistError
 from tests.utils.assertions import (
     assert_equal_track_properties,
     assert_track_dataset_has_tracks,
