+++ tests/utils/builders/track_dataset_provider.py
@@ -1,10 +1,11 @@
 from typing import Literal
 from unittest.mock import Mock
 
+from domain.track_dataset.filtered_track_dataset import FilterByClassTrackDataset
+
 from OTAnalytics.domain.track import Track
 from OTAnalytics.domain.track_dataset.track_dataset import (
     TRACK_GEOMETRY_FACTORY,
-    FilterByClassTrackDataset,
     TrackDataset,
     TrackGeometryDataset,
 )
